@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -58,70 +58,63 @@ import { UpdateBookComponent } from './update-book/update-book.component';
 import { UsersDetailsComponent } from './users-details/users-details.component';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HomeComponent,
-    FooterComponent,
-    SignInComponent,
-    SignUpComponent,
-    BooksComponent,
-    CreateBookComponent,
-    BookDetailsComponent,
-    StudentInfoComponent,
-    DashboardComponent,
-    ProfileComponent,
-    BorrowedBooksComponent,
-    RequestedBooksComponent,
-    BookRecordComponent,
-    EditProfileComponent,
-    ChangePasswordComponent,
-    BooksApprovalComponent,
-    IssuedBooksComponent,
-    BorrowHistoryComponent,
-    UsersDetailsComponent,
-    UpdateBookComponent,
-    BooksDetailsComponent,
-    NavbarComponent,
-    HeroSectionComponent,
-    AboutComponent,
-    KeyFeatureComponent,
-    ContactComponent,
-    ServiceComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    HttpClientModule,
-    MatButtonModule,
-    MatIconModule,
-    MatInputModule,
-    MatFormFieldModule,
-    MatCardModule,
-    FlexLayoutModule,
-    ReactiveFormsModule,
-    FormsModule,
-    MatSelectModule,
-    MatChipsModule,
-    MatPaginatorModule,
-    MatTableModule,
-    MatSortModule,
-    MatListModule,
-    ButtonModule,
-    InputTextModule,
-    PasswordModule,
-    CheckboxModule,
-    RatingModule,
-    FieldsetModule,
-    AutoFocusModule,
-    MatProgressBarModule,
-    DividerModule,
-    InputTextareaModule,
-    FileUploadModule,
-    MatSnackBarModule,
-  ],
-  providers: [],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HomeComponent,
+        FooterComponent,
+        SignInComponent,
+        SignUpComponent,
+        BooksComponent,
+        CreateBookComponent,
+        BookDetailsComponent,
+        StudentInfoComponent,
+        DashboardComponent,
+        ProfileComponent,
+        BorrowedBooksComponent,
+        RequestedBooksComponent,
+        BookRecordComponent,
+        EditProfileComponent,
+        ChangePasswordComponent,
+        BooksApprovalComponent,
+        IssuedBooksComponent,
+        BorrowHistoryComponent,
+        UsersDetailsComponent,
+        UpdateBookComponent,
+        BooksDetailsComponent,
+        NavbarComponent,
+        HeroSectionComponent,
+        AboutComponent,
+        KeyFeatureComponent,
+        ContactComponent,
+        ServiceComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        MatButtonModule,
+        MatIconModule,
+        MatInputModule,
+        MatFormFieldModule,
+        MatCardModule,
+        FlexLayoutModule,
+        ReactiveFormsModule,
+        FormsModule,
+        MatSelectModule,
+        MatChipsModule,
+        MatPaginatorModule,
+        MatTableModule,
+        MatSortModule,
+        MatListModule,
+        ButtonModule,
+        InputTextModule,
+        PasswordModule,
+        CheckboxModule,
+        RatingModule,
+        FieldsetModule,
+        AutoFocusModule,
+        MatProgressBarModule,
+        DividerModule,
+        InputTextareaModule,
+        FileUploadModule,
+        MatSnackBarModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
